Added a player interaction with new key and doors
