+++ ACID/Assets/Models/cyberpunk-datapods/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6678f45036f27c041a5e58f6b99d3f27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
