+++ ACID/Assets/Models/cyberpunk-datapods/Materials/ALBEDO_BLUE.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f291a4bf9bd6dc44cbcc67c38e5159f7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
