+++ ACID/Assets/Models/cyberpunk-datapods/source.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 13d12ff7c312a8a4c8f33bb47e0cf43b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
