+++ ACID/Assets/Models/cyberpunk-datapods/textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee55a32a4aa70474e871b2d96f36dd65
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
