+++ ACID/Assets/Scripts/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6b949efe9f1e2a745aefe776f02b6cc1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
