+++ ACID/Assets/Scripts/Character/Character.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2294c8265d14f4bc186b89a51fd7faf7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
