+++ ACID/Assets/Scripts/Character/Interact.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f89a8e348568b294fb384652d2233124
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
