--- ACID/Assets/Scripts/MouseLook.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2aa5e2806d16643dfaa79f00a8757296
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
