--- ACID/Assets/Scripts/SceneLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3c50c22cb4a5f498681004e25407797e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
