+++ ACID/ProjectSettings/TagManager.asset
@@ -5,7 +5,7 @@ TagManager:
   serializedVersion: 2
   tags:
   - Ground
-  - Ceiling
+  - Key
   layers:
   - Default
   - TransparentFX
@@ -17,7 +17,7 @@ TagManager:
   - 
   - PostProcessing
   - Ground
-  - Ceiling
+  - Interactable
   - 
   - 
   -
